Complete Chat separate component
